Reduce hosted web build memory
