+++ apps/web/src/lib/device-sync/hosted-runtime-authority.ts
@@ -24,11 +24,11 @@ import {
   type HostedExecutionDeviceSyncRuntimeSnapshotResponse,
   type HostedExecutionDeviceSyncRuntimeTokenBundle,
 } from "@murphai/device-syncd/hosted-runtime";
-import {
-  resolveConfiguredDeviceSyncProviderManifest,
-  resolveDeviceProviderMatchKeys,
-} from "@murphai/device-syncd/config";
 import { buildJunctionProviderSourceInstanceKey } from "@murphai/device-syncd/connect-config";
+import {
+  resolveConfiguredDeviceSyncProviderCredentialPolicy,
+} from "@murphai/device-syncd/provider-credential-policy";
+import { resolveDeviceProviderMatchKeys } from "@murphai/device-syncd/provider-match";
 import type { HostedRuntimeRedactedJson } from "@murphai/hosted-execution/runtime-control";
 
 import { createHostedDeviceSyncControlPlane } from "./control-plane";
@@ -1163,8 +1163,7 @@ function validateHostedRuntimeCredentialMutation(input: {
     | Extract<HostedExecutionDeviceSyncRuntimeCredentialUpdate, { clearTokens: true; kind: "oauth_tokens" }>;
   provider: string;
 }): void {
-  const manifest = resolveConfiguredDeviceSyncProviderManifest(input.provider);
-  const policy = manifest?.credentialPolicy;
+  const policy = resolveConfiguredDeviceSyncProviderCredentialPolicy(input.provider);
 
   switch (input.credential.kind) {
     case "oauth_tokens":
