+++ apps/web/test/next-config.test.ts
@@ -16,6 +16,8 @@ import {
   resolveHostedWebDistDir,
 } from "../next-artifacts";
 import {
+  HOSTED_WEB_NEXT_TSCONFIG_PATH,
+  HOSTED_WEB_PRODUCTION_BUILD_CPUS,
   HOSTED_WEB_WORKFLOW_OPTIONS,
   WORKSPACE_SOURCE_PACKAGE_NAMES,
   buildHostedWebClientEnv,
@@ -105,6 +107,40 @@ test("hosted web tsconfig resolves Temporal orchestration-control from source", 
     tsconfig.compilerOptions?.paths?.["@murphai/device-syncd/providers/junction-client"],
     ["packages/device-syncd/src/providers/junction-client.ts"],
   );
+  assert.deepEqual(
+    tsconfig.compilerOptions?.paths?.["@murphai/device-syncd/provider-credential-policy"],
+    ["packages/device-syncd/src/provider-credential-policy.ts"],
+  );
+  assert.deepEqual(
+    tsconfig.compilerOptions?.paths?.["@murphai/device-syncd/provider-match"],
+    ["packages/device-syncd/src/provider-match.ts"],
+  );
+  assert.deepEqual(
+    tsconfig.compilerOptions?.paths?.["@murphai/device-syncd/providers/junction-config"],
+    ["packages/device-syncd/src/providers/junction-config.ts"],
+  );
+});
+
+test("hosted web build tsconfig keeps tests out of Next production checks", () => {
+  const tsconfig = JSON.parse(
+    readFileSync(path.join(repoRoot, "apps/web/tsconfig.next.json"), "utf8"),
+  ) as {
+    extends?: string;
+    include?: readonly string[];
+    exclude?: readonly string[];
+    compilerOptions?: {
+      tsBuildInfoFile?: string;
+    };
+  };
+
+  assert.equal(tsconfig.extends, "./tsconfig.json");
+  assert.equal(tsconfig.compilerOptions?.tsBuildInfoFile, ".next/cache/tsconfig.next.tsbuildinfo");
+  assert.ok(tsconfig.include?.includes("app/**/*.tsx"));
+  assert.ok(tsconfig.include?.includes("src/**/*.ts"));
+  assert.ok(tsconfig.include?.includes(".next/types/**/*.ts"));
+  assert.ok(!tsconfig.include?.includes("test/**/*.ts"));
+  assert.ok(!tsconfig.include?.includes("test/**/*.tsx"));
+  assert.ok(tsconfig.exclude?.includes("test"));
 });
 
 test("hosted web dist-dir selection reserves a dedicated artifact directory for interactive dev", () => {
@@ -249,7 +285,10 @@ test("hosted web dev filesystem cache defaults off and allows explicit opt-in", 
 test("next.config keeps Turbopack focused on the repo root without custom workspace rewrite rules", () => {
   assert.equal(productionNextConfig.turbopack?.root, process.cwd());
   assert.equal(productionNextConfig.webpack, undefined);
-  assert.equal(productionNextConfig.typescript, undefined);
+  assert.deepEqual(productionNextConfig.typescript, {
+    tsconfigPath: HOSTED_WEB_NEXT_TSCONFIG_PATH,
+  });
+  assert.equal(productionNextConfig.experimental?.cpus, HOSTED_WEB_PRODUCTION_BUILD_CPUS);
 });
 
 test("hosted runtime issue imports avoid the runtime-state Node barrel", () => {
