+++ apps/web/tsconfig.json
@@ -232,8 +232,17 @@
       "@murphai/device-syncd/provider-label": [
         "packages/device-syncd/src/provider-label.ts"
       ],
+      "@murphai/device-syncd/provider-credential-policy": [
+        "packages/device-syncd/src/provider-credential-policy.ts"
+      ],
+      "@murphai/device-syncd/provider-match": [
+        "packages/device-syncd/src/provider-match.ts"
+      ],
       "@murphai/device-syncd/providers/junction-client": [
         "packages/device-syncd/src/providers/junction-client.ts"
+      ],
+      "@murphai/device-syncd/providers/junction-config": [
+        "packages/device-syncd/src/providers/junction-config.ts"
       ],
       "@murphai/device-syncd/providers/oura": [
         "packages/device-syncd/src/providers/oura.ts"
